Mark embeddings on VectorQuantizerEMA as non-trainable.
This variable is updated as part of the forward pass (during training) and is
not updated by the optimiser. In TF `trainable` is used to indicate variables
that should be managed by an optimizer, so even though we update this during
training and it is a parameter of our model, we should set `trainable=False`.

Fixes #209.

PiperOrigin-RevId: 385501742